Add es2020 transformation (#35518)
* Add es2020 transformation

* Add missing comma
