+++ src/compiler/tsconfig.json
@@ -43,6 +43,7 @@
         "transformers/es2017.ts",
         "transformers/es2018.ts",
         "transformers/es2019.ts",
+        "transformers/es2020.ts",
         "transformers/esnext.ts",
         "transformers/jsx.ts",
         "transformers/es2016.ts",
