+++ tests/baselines/reference/nullishCoalescingOperator_es2020.types
@@ -1,4 +1,4 @@
-=== tests/cases/conformance/expressions/nullishCoalescingOperator/nullishCoalescingOperator_esnext.ts ===
+=== tests/cases/conformance/expressions/nullishCoalescingOperator/nullishCoalescingOperator_es2020.ts ===
 declare const a1: 'literal' | undefined | null
 >a1 : "literal" | null | undefined
 >null : null
