+++ tests/cases/conformance/expressions/nullishCoalescingOperator/nullishCoalescingOperator_es2020.ts
@@ -1,5 +1,5 @@
 // @strict: true
-// @target: esnext
+// @target: es2020
 
 declare const a1: 'literal' | undefined | null
 declare const a2: '' | undefined | null
